+++ algorithms/median_filter/openmp/algorithm_median_filter_openmp.cpp
@@ -33,8 +33,8 @@ namespace algorithms
 		void Algorithm::compute3x3()
 		{
 			Frame result = _frame.clone();
-			const int nRows = result.nRows;
-			const int nCols = result.nCols;
+			const int nRows = static_cast<int>(result.nRows);
+			const int nCols = static_cast<int>(result.nCols);
 
 #pragma omp parallel for
 			for (int i = 1; i < nRows - 1; i++)
@@ -50,9 +50,9 @@ namespace algorithms
 
 		void Algorithm::median3x3(int x, int y, const Frame& frame, Frame& result, int indexRes)
 		{
-			const int maskSize = 9;
+			const size_t maskSize = 9;
 			/*Indexes from original frame for mask*/
-			int indexes[maskSize] = { (y - 1)*frame.nCols + x - 1, (y - 1)*frame.nCols + x, (y - 1)*frame.nCols + x + 1,
+			size_t indexes[maskSize] = { (y - 1)*frame.nCols + x - 1, (y - 1)*frame.nCols + x, (y - 1)*frame.nCols + x + 1,
 							  y*frame.nCols + x - 1, y*frame.nCols + x, y*frame.nCols + x + 1,
 							  (y + 1)*frame.nCols + x - 1, (y + 1)*frame.nCols + x, (y + 1)*frame.nCols + x + 1 };
 
@@ -80,9 +80,9 @@ namespace algorithms
 
 		void Algorithm::median5x5(int x, int y, const Frame& frame, Frame& result, int indexRes)
 		{
-			const int maskSize = 25;
+			const size_t maskSize = 25;
 			/*Indexes from original frame for mask*/
-			int indexes[maskSize] = { (y - 2)*frame.nCols + x - 2,
+			size_t indexes[maskSize] = { (y - 2)*frame.nCols + x - 2,
 								(y - 2)*frame.nCols + x - 1,
 								(y - 2)*frame.nCols + x,
 								(y - 2)*frame.nCols + x + 1,
@@ -133,8 +133,8 @@ namespace algorithms
 		void Algorithm::compute5x5()
 		{
 			Frame result = _frame.clone();
-			const int nRows = result.nRows;
-			const int nCols = result.nCols;
+			const int nRows = static_cast<int>(result.nRows);
+			const int nCols = static_cast<int>(result.nCols);
 
 #pragma omp parallel for
 			for (int i = 2; i < nRows - 2; i++)
